remote-access-path: Support stack variables in "virtual" registers
While the original access-path library modeled local variables as
either being on the stack (at a memory location) or in a register, the
lldb-based debugger nub can not always provide this level of
specificity. For this reason, for stack variable values without an
address, it returns a virtual register number with a negative
index. This change to the remote-access-path accomodates this behavior
in the nub by no longer insisting that the register index be a known
one when it is negative.

* sources/runtime-manager/remote-access-path/stack-frames.dylan
  (read-frame-lexicals): Handle variable locations with a negative
   register index by creating a transient <active-remote-register>
   in the #"virtual" register category.