run-time: Allow the debugger to hold newly-created threads
When the debugger calls spy-create-application-thread, it needs to
stop the created thread before it executes a hard breakpoint (which
would interfere with the execution of the spy call). This change adds
a spy function that the debugger can call to force newly-created
threads to wait until the debugger can detect and control them.

* sources/lib/run-time/llvm-posix-threads.c
  (thread_hold): New flag variable.
  (thread_hold_lock): New mutex variable.
  (thread_hold_change): New condition variable.
  (dylan_thread_trampoline): Wait to execute the Dylan thread function
   until thread_hold is no longer set.
  (spy_hold_threads): New function to be used by the debugger to set
   or clear the thread_hold state.

* sources/lib/run-time/llvm-runtime.h
  (spy_hold_threads): New declaration.