run-time: Add a spy interface for retrieving signals to ignore
The debugger nub needs to know what Unix signals are utilized by the
Dylan run-time (including the garbage collector) so that it can avoid
stopping the process when they are raised. This change provides a spy
interface that the nub can use at process initialization time to query
for the set of signals to exclude.

* sources/lib/run-time/spy-interfaces.h
  (SPY_MAX_RUNTIME_SIGNALS, spy_get_runtime_signal): New declarations.

* sources/lib/run-time/unix-spy-interfaces.c
  (runtime_signals, runtime_signals_count): New static definitions.
  (spy_get_runtime_signal): New spy function for enumerating
   signals used by the garbage collector and run-time support.