system: Fix open function call punning
The open(2) function uses varargs, so that when the O_CREAT flag is
passed in the second argument then a third file mode option is
used. While many ABIs are fine with always passing three arguments to
open(), the ABI used on Darwin 64-bit Arm platforms can not easily
handle this. This necessitates adding a C wrapper for the call to
open().

* sources/system/file-system/unix-interface.dylan
  (unix-open): Change to use the system_open wrapper function.

* sources/system/unix-portability.c
  (system_open): New wrapper function that calls open() in accordance
   with its varargs declaration.